# docs/swagger.yaml
definitions:
  entity.ConfigSpiderData:
    properties:
      cmd:
        description: 自定义爬虫
        type: string
      col:
        type: string
      display_name:
        type: string
      engine:
        description: 可配置爬虫
        type: string
      name:
        description: 通用
        type: string
      remark:
        type: string
      settings:
        type: object
      stages:
        items:
          $ref: '#/definitions/entity.Stage'
        type: array
      start_stage:
        type: string
      start_url:
        type: string
      type:
        type: string
    type: object
  entity.Field:
    properties:
      attr:
        type: string
      css:
        type: string
      name:
        type: string
      next_stage:
        type: string
      remark:
        type: string
      xpath:
        type: string
    type: object
...
  entity.Stage:
    properties:
      fields:
        items:
          $ref: '#/definitions/entity.Field'
        type: array
      is_list:
        type: boolean
      list_css:
        type: string
      list_xpath:
        type: string
      name:
        type: string
      page_attr:
        type: string
      page_css:
        type: string
      page_xpath:
        type: string
    type: object
  model.Env:
    properties:
      name:
        type: string
      value:
        type: string
    type: object
  model.Node:
    properties:
      _id:
        type: string
      create_ts:
        type: string
      description:
        type: string
      hostname:
        type: string
      ip:
        type: string
      is_master:
        description: 前端展示
        type: boolean
      key:
        description: 用于唯一标识节点，可能是mac地址，可能是ip地址
        type: string
      mac:
        type: string
      name:
        type: string
      port:
        type: string
      status:
        type: string
      update_ts:
        type: string
      update_ts_unix:
        type: integer
    type: object
  model.Project:
    properties:
      _id:
        type: string
      create_ts:
        type: string
      description:
        type: string
      name:
        type: string
      spiders:
        description: 前端展示
        items:
          $ref: '#/definitions/model.Spider'
        type: array
      tags:
        items:
          type: string
        type: array
      update_ts:
        type: string
      user_id:
        type: string
      username:
        type: string
    type: object
  model.Schedule:
    properties:
      _id:
        type: string
      create_ts:
        type: string
      cron:
        type: string
      description:
        type: string
      enabled:
        type: boolean
      entry_id:
        type: integer
      message:
        type: string
      name:
        type: string
      node_ids:
        items:
          type: string
        type: array
      nodes:
        items:
          $ref: '#/definitions/model.Node'
        type: array
      param:
        type: string
      run_type:
        type: string
      scrapy_log_level:
        type: string
      scrapy_spider:
        type: string
      spider_id:
        type: string
      spider_name:
        description: 前端展示
        type: string
      status:
        type: string
      update_ts:
        type: string
      user_id:
        type: string
      user_name:
        type: string
    type: object
  model.Spider:
    properties:
      _id:
        description: 爬虫ID
        type: string
      cmd:
        description: 自定义爬虫
        type: string
      col:
        description: 结果储存位置
        type: string
      config:
        $ref: '#/definitions/entity.ConfigSpiderData'
        description: 可配置爬虫配置
        type: object
      create_ts:
        type: string
      dedup_field:
        description: 去重字段
        type: string
      dedup_method:
        description: 去重方式
        type: string
      display_name:
        description: 爬虫显示名称
        type: string
      envs:
        description: 环境变量
        items:
          $ref: '#/definitions/model.Env'
        type: array
      file_id:
        description: GridFS文件ID
        type: string
      git_auto_sync:
        description: Git 是否自动同步
        type: boolean
      git_branch:
        description: Git 分支
        type: string
      git_has_credential:
        description: Git 是否加密
        type: boolean
      git_password:
        description: Git 密码
        type: string
      git_sync_error:
        description: Git 同步错误
        type: string
      git_sync_frequency:
        description: Git 同步频率
        type: string
      git_url:
        description: Git URL
        type: string
      git_username:
        description: Git 用户名
        type: string
      is_dedup:
        description: 去重
        type: boolean
      is_git:
        description: Git 设置
        type: boolean
      is_long_task:
        description: 长任务
        type: boolean
      is_public:
        description: 是否公开
        type: boolean
      is_scrapy:
        description: Scrapy 爬虫（属于自定义爬虫）
        type: boolean
      is_web_hook:
        description: Web Hook
        type: boolean
      last_run_ts:
        description: 前端展示
        type: string
      last_status:
        description: 最后执行状态
        type: string
      latest_tasks:
        description: 最近任务列表
        items:
          $ref: '#/definitions/model.Task'
        type: array
      name:
        description: 爬虫名称（唯一）
        type: string
      project_id:
        description: 项目ID
        type: string
      remark:
        description: 备注
        type: string
      site:
        description: 爬虫网站
        type: string
      spider_names:
        description: 爬虫名称列表
        items:
          type: string
        type: array
      src:
        description: 源码位置
        type: string
      template:
        description: 可配置爬虫
        type: string
      type:
        description: 爬虫类别
        type: string
      update_ts:
        type: string
      user_id:
        description: 时间
        type: string
      username:
        description: 用户名称
        type: string
      web_hook_url:
        description: Web Hook URL
        type: string
    type: object
  model.Task:
    properties:
      _id:
        type: string
      cmd:
        type: string
      create_ts:
        type: string
      error:
        type: string
      error_log_count:
        type: integer
      finish_ts:
        type: string
      log_path:
        type: string
      node_id:
        type: string
      node_name:
        type: string
      param:
        type: string
      pid:
        type: integer
      result_count:
        type: integer
      run_type:
        type: string
      runtime_duration:
        type: number
      schedule_id:
        type: string
      spider_id:
        type: string
      spider_name:
        description: 前端数据
        type: string
      start_ts:
        type: string
      status:
        type: string
      total_duration:
        type: number
      update_ts:
        type: string
      user_id:
        type: string
      username:
        type: string
      wait_duration:
        type: number
    type: object
...
info:
  contact: {}
  license: {}
paths:
...
  /projects:
    get:
      description: Get projects
      parameters:
        - description: With the bearer started
          in: header
          name: Authorization
          required: true
          type: string
        - description: projects
          in: query
          name: tag
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get projects
      tags:
        - project
    put:
      consumes:
        - application/json
      description: Put project
      parameters:
        - description: With the bearer started
          in: header
          name: Authorization
          required: true
          type: string
        - description: post project
          in: body
          name: p
          required: true
          schema:
            $ref: '#/definitions/model.Project'
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Put project
      tags:
        - project
  /projects/{id}:
    delete:
      description: Delete project
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: project id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Delete project
      tags:
        - project
    post:
      consumes:
        - application/json
      description: Post project
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: project id
          in: path
          name: id
          required: true
          type: string
        - description: project item
          in: body
          name: item
          required: true
          schema:
            $ref: '#/definitions/model.Project'
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Post project
      tags:
        - project
  /projects/tags:
    get:
      description: Get projects tags
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get project tags
      tags:
        - project
...
  /schedules:
    get:
      description: Get spider list
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: page num
          in: query
          name: page_num
          type: string
        - description: page size
          in: query
          name: page_size
          type: string
        - description: keyword
          in: query
          name: keyword
          type: string
        - description: project_id
          in: query
          name: project_id
          type: string
        - description: type
          in: query
          name: type
          type: string
        - description: sort_key
          in: query
          name: sort_key
          type: string
        - description: sort_direction
          in: query
          name: sort_direction
          type: string
        - description: owner_type
          in: query
          name: owner_type
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get spider list
      tags:
        - spider
    put:
      consumes:
        - application/json
      description: Put schedule
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: schedule item
          in: body
          name: item
          required: true
          schema:
            $ref: '#/definitions/model.Schedule'
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Put schedule
      tags:
        - schedule
  /schedules/{id}:
    delete:
      description: Delete schedule
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: schedule id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Delete schedule
      tags:
        - schedule
    get:
      description: Get schedule  by id
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: schedule id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "400":
          description: Bad Request
          schema:
            type: json
      summary: Get schedule by id
      tags:
        - schedule
    post:
      consumes:
        - application/json
      description: Post schedule
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: schedule id
          in: path
          name: id
          required: true
          type: string
        - description: schedule item
          in: body
          name: newItem
          required: true
          schema:
            $ref: '#/definitions/model.Schedule'
            type: object
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: Post schedule
      tags:
        - schedule
  /schedules/{id}/disable:
    post:
      consumes:
        - application/json
      description: disable schedule
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: schedule id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: disable schedule
      tags:
        - schedule
  /schedules/{id}/enable:
    post:
      consumes:
        - application/json
      description: enable schedule
      parameters:
        - description: Authorization token
          in: header
          name: Authorization
          required: true
          type: string
        - description: schedule id
          in: path
          name: id
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            type: json
        "500":
          description: Internal Server Error
          schema:
            type: json
      summary: enable schedule
      tags:
        - schedule
...
swagger: "2.0"